played around with the scripts
I actuall forgot what I have done as I was busy with my projects at my
work. As far as I recall I played around with the parametric and non
parametric fitting. I was not satisfied on how to determine the goodness
of the fit (deviance, chisquare or montecarlo simulations).  But finally
used the code in pfitindivudial_bo2010_sharpness.m to find the
thresholds for the two best performing blind particpants.  I have sent
he results in a draft document to professor so that include them in the
manuscript. For now that is it. Will clean up the scripts in future
commits. :)